// src/openapi/client/models/ConfigCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ConfigCreate = {
    phones: Array<string>;
    emails: Array<string>;
    supportPhone?: (string | null);
    supportEmail?: (string | null);
    regularShippingPrice: (number | string);
    expressShippingPrice: (number | string);
};